feat: add admin access handling logic
