+++ src/main/kotlin/jsonweb/exitserver/common/BaseTimeEntity.kt
